requirements.txt
telethon==1.36.0
APScheduler==3.10.4
python-dotenv==1.0.1
aiogram==3.13.1
